refactor(mcp-apps): simplify kody widget imports
